Julian Date in common functions
